[REFACTOR] undo relocation of commons classes (Auditable, EntityId)
