+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/employers/application/EmployerRetriever.kt
@@ -3,9 +3,9 @@ package uk.gov.justice.digital.hmpps.jobsboard.api.employers.application
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.employers.domain.Employer
 import uk.gov.justice.digital.hmpps.jobsboard.api.employers.domain.EmployerRepository
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 
 @Service
 class EmployerRetriever(
