+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/application/JobRetriever.kt
@@ -3,7 +3,7 @@ package uk.gov.justice.digital.hmpps.jobsboard.api.jobs.application
 import org.springframework.data.domain.Page
 import org.springframework.data.domain.Pageable
 import org.springframework.stereotype.Service
-import uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain.EntityId
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain.Job
 import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain.JobRepository
