+++ src/main/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/domain/Auditable.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain
+package uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain
 
 import jakarta.persistence.Column
 import jakarta.persistence.EntityListeners
