+++ src/test/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/employers/application/EmployerCreatorShould.kt
@@ -10,8 +10,8 @@ import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.employers.domain.Employer
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.application.TestBase
 import java.util.*
 import kotlin.test.Test
