+++ src/test/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/employers/domain/EntityIdShould.kt
@@ -1,6 +1,7 @@
-package uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain
+package uk.gov.justice.digital.hmpps.jobsboard.api.employers.domain
 
 import org.junit.jupiter.api.Test
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
