+++ src/test/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/application/JobCreatorShould.kt
@@ -11,7 +11,7 @@ import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.never
 import org.mockito.kotlin.verify
 import org.mockito.kotlin.whenever
-import uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain.EntityId
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain.Job
 import java.time.LocalDate
 import java.time.Month.JULY
