+++ src/test/kotlin/uk/gov/justice/digital/hmpps/jobsboard/api/jobs/application/JobRetrieverShould.kt
@@ -6,7 +6,7 @@ import org.mockito.Mockito.times
 import org.mockito.Mockito.`when`
 import org.mockito.junit.jupiter.MockitoExtension
 import org.mockito.kotlin.verify
-import uk.gov.justice.digital.hmpps.jobsboard.api.commons.domain.EntityId
+import uk.gov.justice.digital.hmpps.jobsboard.api.entity.EntityId
 import uk.gov.justice.digital.hmpps.jobsboard.api.jobs.domain.Job
 import java.util.*
 import kotlin.test.Test
